feat: Refactor folder preview grid item display (test-folder-anim)
This commit refactors how folder preview grid items are displayed and manages
constants related to folder grid dimensions.

Key changes include:

- **Constants Relocation:**
  - `FOLDER_PREVIEW_COLUMNS` and `FOLDER_PREVIEW_ROWS` constants have been
    moved from `HomeUtil.kt` to `GridUtil.kt` for better organization.

- **Preview Grid Item Display:**
  - The `PreviewFolderGridLayout` now iterates directly over `gridItems`
    instead of taking a fixed number of items. This ensures that all
    available preview items are considered for display.
  - The logic for calculating the position and size of each preview grid
    item has been adjusted to work with the direct iteration.

- **Folder Grid Item Filtering:**
  - In `GetPreviewFolderGridItemsUseCase`, the preview grid items are now
    filtered based on the calculated `columns` and `rows` derived from
    `getGridDimension`. This ensures that only the relevant number of
    items for the preview are returned.

These changes improve the accuracy and consistency of folder preview
display and consolidate related constants.
